Make the project buildable again & enable multiplatform tests (#202)
For context: I'm the maintainer of
[pyHanko](https://github.com/MatthiasValvekens/pyHanko), which has had
an optional dependency on python-pkcs11 for PKCS#11 support for 4+ years
at this point.

I was led down this rabbit hole by a user report complaining about the
fact that the current latest release (from 2020, prior to this project
moving to the pyauth org) isn't installable on Windows with Python 3.12
or greater (MatthiasValvekens/pyHanko#558, #197).

While I'm not much of a Windows dev, this issue seemed fairly obvious to
fix, so I set out to do so, only to discover that the isolation enforced
by the typical modern Python build flow (in particular, when running `uv
build`) is fundamentally at odds with the way this project was
historically structured/built. So besides fixing #197, I also addressed
a number of other issues:

- Restructured the Cython build to actually work when executed from a
build frontend (this included fixing the source distribution post-`uv`
migration)
- Refactored the Cython layer to remove deprecated constructs
(conditional compilation, Cython-level `include`s, ...).
- The only platform-dependent logic was in the code that loads the
PKCS#11 library, so I moved that to a `load_module.c` file that exposes
a common C interface with platform-dependent implementations.
- I had to consolidate the Cython files a bit to be compliant with the
"include-free" style while still building everything as a single C
extension.
- Eliminated the dependency on `oscrypto` in the tests, which in turn
gets rid of the hard dependency on OpenSSL 1.1.x, which in turn also
means that all the hacks in the GitHub workflow for tests could be
removed as well.

I'm upstreaming my changes in the hope that it'll get this project into
a state where new releases can be cut more easily again, since I rely on
this library quite heavily. I'd like to avoid having to fork it if
that's at all avoidable...


-----

PS: Unsolicited opinion about testing: For a project like this, IMO
there's also no way around setting up a multiplatform test matrix
(including at least one linux/windows/macOS variant each). Ideally that
would include testing against actual HSMs, but operating a hardware test
bench isn't exactly trivial or cheap for an OSS project, so the fact
that this isn't being done is certainly understandable. Funnily enough
the readme still claims that there is CI running against real hardware,
but that hasn't been the case for several years at this point ;). I have
several devices lying around at home that I can test with, but I'm
currently limited to manual testing on that front.


PPS: With these changes, the `cibuildwheel` action appears to work, by
the way.

---

EDIT: I added `windows-latest` and `macos-latest` to the CI test matrix
now :).